Fichier fonctionne

+++ Affichage/Console.cpp
@@ -1,23 +1,24 @@
 #include "Console.hpp"
-#include "../Sauvegarde/SauvegardeTxt.hpp"
-
-extern std::string cheminDossierSortie; // Déclaration de la variable globale
 
 Console::Console() {}
 
 Console::~Console() {}
 
-void Console::run(Grille *g) {
-    std::string cheminDossier = cheminDossierSortie; // Utiliser le chemin du dossier de sortie
-    for (const auto& fichier : std::filesystem::directory_iterator(cheminDossier)) { // Parcours tout les éléments
-        if (std::filesystem::is_regular_file(fichier.path())) { // .path() : Récupère le chemin complet || is_regular_file() : Vérifie si cet élément est un fichier.
-            std::filesystem::remove(fichier.path()); // Supprime le fichier
+void Console::run(Grille *g, Fichier *f) {
+    Sauvegarde *s = new SauvegardeFichierTxt();
+    std::string cheminDossier = f->getDossierSortie(); // Utiliser le chemin du dossier de sortie
+    std::cout << f->getDossierSortie() << std::endl;
+    if (!std::filesystem::is_empty(cheminDossier)) {
+        std::cout << "Done" << std::endl;
+        for (const auto& fichier : std::filesystem::directory_iterator(cheminDossier)) { // Parcours tout les éléments
+            if (std::filesystem::is_regular_file(fichier.path())) { // .path() : Récupère le chemin complet || is_regular_file() : Vérifie si cet élément est un fichier.
+                std::filesystem::remove(fichier.path()); // Supprime le fichier
+                std::cout << "Fichier supprimé" << std::endl;
+            } 
         }
     }
 
     int iterations;
-    Sauvegarde *s;
-    s = new SauvegardeFichierTxt();
     std::cout << "Nombre de cycles du jeu de la vie : ";
     std::cin >> iterations;
     for (int cycle = 0; cycle < iterations; cycle++) {
